samples: add verilog integration sample
## What has been done

- Added a sample that shows how to integrate verilog source with your firmware application
- Demonstrate how to run formal verification as part of build process

This example demonstrates how a safety critical logical part of an application can be developed in verilog and included as C code. This is accomplished not using verilator but rather using CXXRTL generator that comes with yosys. The code that it generates is very simple and basically just implements the verilog logic without much extra stuff.

## Proof

- see samples/hdl/verilog

## Known issues/limitations

* Currently requires OSS CAD suite (included in swedish embedded docker image)

Closes #73